# packages/api-client/spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/google": {
      "post": {
        "description": "Verifies the Google ID token, links the provider account by verified email (no duplicate accounts), blocks under-18 and suspended accounts, and issues a token pair.",
        "operationId": "SocialController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string"
                  },
                  "deviceName": {
                    "type": "string"
                  }
                },
                "required": [
                  "idToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair; isNewUser=true flags first-time users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "handle": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        },
                        "avatarUrl": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "isNewUser": {
                      "type": "boolean"
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "accessTokenExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "refreshTokenExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "INVALID_ID_TOKEN | EMAIL_NOT_VERIFIED"
          },
          "403": {
            "description": "UNDERAGE_BLOCK | ACCOUNT_SUSPENDED | ACCOUNT_DELETED"
          }
        },
        "summary": "Sign in / sign up with Google",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/apple": {
      "post": {
        "description": "Verifies the Apple identity token against Apple's JWKS, links by email when present, blocks under-18 and suspended accounts, and issues a token pair.",
        "operationId": "SocialController_apple",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idToken": {
                    "type": "string"
                  },
                  "deviceName": {
                    "type": "string"
                  }
                },
                "required": [
                  "idToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair; isNewUser=true flags first-time users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "handle": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "nullable": true
                        },
                        "avatarUrl": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "isNewUser": {
                      "type": "boolean"
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "accessTokenExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "refreshTokenExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "INVALID_ID_TOKEN"
          },
          "403": {
            "description": "UNDERAGE_BLOCK | ACCOUNT_SUSPENDED | ACCOUNT_DELETED"
          }
        },
        "summary": "Sign in / sign up with Apple",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moments API",
    "description": "Moments REST API — Phase 0 skeleton (F5/F6)",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}